Clean Up Number 2
Deleted everything in the unused folder because the folder structure was
not kept intact.  I have no idea what goes where to restore it even if I
wanted to.  I also made sure that the icon assignments had the full path
instead of just the filename.  Hopefully, there should be better compile
times.